% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forwardfeatureselection.R
\name{forwardfeatureselection}
\alias{forwardfeatureselection}
\title{Forward Feature Selection}
\usage{
forwardfeatureselection(
  model,
  training,
  test,
  featurelist,
  includePlot = FALSE
)
}
\arguments{
\item{model}{The ML models used to classify the data, typically SVM with a given kernel}

\item{training}{Training dataset as a data.frame with classification column and column for each feature.}

\item{test}{Test dataset with matching columns to training.}

\item{featurelist}{List of features to order}

\item{includePlot}{Show number of features vs accuracy line plot (default:FALSE)}
}
\value{
Ordered list of most discriminative features when classifying the dataset along with training and test accuracy, sensitivity and specificity
}
\description{
This function performs forward feature selection on the given list of features, placing them in order of discriminative power using a given model on the given dataset up to the accuracy plateau.
}
\examples{
forwardfeatureselection(linear_svm,data_train,data_test,subset(colnames(data_train),select=-c(classification)))
}
\keyword{feature}
\keyword{forward}
\keyword{selection}
