% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameters.R
\name{selectsvmkernel}
\alias{selectsvmkernel}
\title{Tuning SVM kernel
This function trains SVMs with a range of kernels (linear, polynomial degree 2, 3 and 4, radial and sigmoid) using cross validation so the optimal kernel can be chosen (using the resulting plots). If specified (by showplots=F) the plots are saved as jpegs.}
\usage{
selectsvmkernel(data, title, showplots = T, output_prefix = "")
}
\arguments{
\item{data}{Dataset: dataframe containing classification column and all other column features. Both the training and test datasets will be taken from this dataset.}

\item{title}{Title to be used for the resulting boxplot}
}
\value{
Dataframe containing test and training accuracy, sensitivity and specificity
}
\description{
Tuning SVM kernel
This function trains SVMs with a range of kernels (linear, polynomial degree 2, 3 and 4, radial and sigmoid) using cross validation so the optimal kernel can be chosen (using the resulting plots). If specified (by showplots=F) the plots are saved as jpegs.
}
\examples{
selectsvmkernel(data,'Boxplots showing distribution of training and test performance for random forest using number of trees 1 to 100')
}
\keyword{SVM}
\keyword{hyperparameter}
\keyword{kernel}
\keyword{tuning}
