% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{svmpolynomial3}
\alias{svmpolynomial3}
\title{Polynomial degree 3 SVM
This function implements a polynomial degree 3 SVM using the general svm function (for ease of use in feature selection)}
\usage{
svmpolynomial3(data_train, data_test, includeplot = FALSE)
}
\arguments{
\item{data_train}{Training set: dataframe containing classification column and all other columns features. This is the dataset on which the decision tree model is trained.}

\item{data_test}{Test set: dataframe containing classification column and all other columns features. This is the dataset on which the decision tree model in tested.}

\item{includeplot}{Show performance scatter plot (default:FALSE)}
}
\value{
List containing performance percentages, accessed using training (training accuracy), test (test accuracy), trainsensitivity, testsensitivity, trainspecificity, testspecificity.
}
\description{
Polynomial degree 3 SVM
This function implements a polynomial degree 3 SVM using the general svm function (for ease of use in feature selection)
}
\examples{
svmpolynomial3(data_train,data_test)
}
\keyword{classification}
\keyword{svm}
