% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{svmradial}
\alias{svmradial}
\title{Radial SVM
This function implements a radial SVM using the general svm function (for ease of use in feature selection)}
\usage{
svmradial(data_train, data_test, includeplot = FALSE)
}
\arguments{
\item{data_train}{Training set: dataframe containing classification column and all other columns features. This is the dataset on which the decision tree model is trained.}

\item{data_test}{Test set: dataframe containing classification column and all other columns features. This is the dataset on which the decision tree model in tested.}

\item{includeplot}{Show performance scatter plot (default:FALSE)}
}
\value{
List containing performance percentages, accessed using training (training accuracy), test (test accuracy), trainsensitivity, testsensitivity, trainspecificity, testspecificity.
}
\description{
Radial SVM
This function implements a radial SVM using the general svm function (for ease of use in feature selection)
}
\examples{
"svmradial(data_train,data_test)"
}
\keyword{classification}
\keyword{svm}
