% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{svmsigmoid}
\alias{svmsigmoid}
\title{Sigmoid SVM
This function implements a sigmoid SVM using general svm function (for ease of use in feature selection)}
\usage{
svmsigmoid(data_train, data_test, includeplot = FALSE)
}
\arguments{
\item{data_train}{Training set: dataframe containing classification column and all other columns features. This is the dataset on which the decision tree model is trained.}

\item{data_test}{Test set: dataframe containing classification column and all other columns features. This is the dataset on which the decision tree model in tested.}

\item{includeplot}{Show performance scatter plot (default:FALSE)}
}
\value{
List containing performance percentages, accessed using training (training accuracy), test (test accuracy), trainsensitivity, testsensitivity, trainspecificity, testspecificity.
}
\description{
Sigmoid SVM
This function implements a sigmoid SVM using general svm function (for ease of use in feature selection)
}
\examples{
"svmsigmoid(data_train,data_test)"
}
\keyword{classification}
\keyword{svm}
