% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eGA.R
\name{eGA}
\alias{eGA}
\title{Embryonic Genetic Algorithm
This function is based on Embryonic Genetic Algorithms. It performs feature selection by maintaining an ongoing set of 'good' set of features which are improved run by run. It outputs training and test accuracy, sensitivity and specificity and a list of <=k features.}
\usage{
eGA(
  k = 30,
  data_train,
  data_test,
  mutprob = 0.05,
  includePlot = FALSE,
  maxnumruns = 50
)
}
\arguments{
\item{k}{Maximum number of features in the output feature set (default:30)}

\item{data_train}{Training set: dataframe containing classification column and all other columns features. This is the dataset on which the decision tree model is trained.}

\item{data_test}{Test set: dataframe containing classification column and all other columns features. This is the dataset on which the decision tree model in tested.}

\item{mutprob}{Probability that mutation will be performed for each produced feature set from forward feature selection (default:0.05)}

\item{includePlot}{Show performance scatter plot (default:FALSE)}

\item{maxnumruns}{Maximum number of iterations after which the feature set will be output, if no other termination conditions have been met (default:50)}
}
\value{
List containing (ordered list of) selected features, performance percentages, accessed using training (training accuracy), test (test accuracy), trainsensitivity, testsensitivity, trainspecificity, testspecificity. Also accessed using listofongoing is a list containing the length of the ongoing set at each stage.
}
\description{
Embryonic Genetic Algorithm
This function is based on Embryonic Genetic Algorithms. It performs feature selection by maintaining an ongoing set of 'good' set of features which are improved run by run. It outputs training and test accuracy, sensitivity and specificity and a list of <=k features.
}
\examples{
"eGA(30,data_train,data_test,maxnumruns=100)"
}
\keyword{eGA}
\keyword{embryonic}
\keyword{feature}
\keyword{genetic}
\keyword{selection}
