% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameters.R
\name{selectrfnumtrees}
\alias{selectrfnumtrees}
\title{Tuning number of trees hyperparameter
This function trains random forests with a range of number of trees so the optimal number can be identified (using the resulting plot) with cross validation}
\usage{
selectrfnumtrees(data, maxnum = 100, title = "", output_prefix = "")
}
\arguments{
\item{data}{Dataset: dataframe containing classification column and all other column features. Both the training and test datasets will be taken from this dataset.}

\item{maxnum}{Maximum number of trees to be considered. All numbers between 1 and maxnum will be considered. Default: 100.}

\item{title}{Title to be used for the resulting boxplot}
}
\value{
Dataframe containing test and training accuracy, sensitivity and specificity
}
\description{
Tuning number of trees hyperparameter
This function trains random forests with a range of number of trees so the optimal number can be identified (using the resulting plot) with cross validation
}
\examples{
selectrfnumtrees(data,100,'Boxplots showing distribution of training and test performance for random forest using number of trees 1 to 100')
}
\keyword{forest}
\keyword{hyperparameter}
\keyword{random}
\keyword{tuning}
